Conflicts:
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/Father.cpp
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/GameplayState.cpp
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/MainMenuState.cpp
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/Son.cpp
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/TileSystem.cpp
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/TileSystem.h

+++ 1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/AutoLockingDoor.cpp
@@ -61,7 +61,7 @@ int	AutoLockingDoor::GetType(void)	const
 }
 SGD::Rectangle AutoLockingDoor::GetRect(void) const
 {
-	return AnimationSystem::GetInstance()->GetRect(m_Timestamp, (int)m_ptPosition.x, (int)m_ptPosition.y, SGD::Size{ 1.0f, 1.0f });
+	return AnimationSystem::GetInstance()->GetRect(m_Timestamp, m_ptPosition.x, m_ptPosition.y, SGD::Size{ 1.0f, 1.0f });
 }
